third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/the-form-element/form-action-reflection-with-base-url-expected.txt

This is a testharness.js-based test.
PASS An action URL should be resolved relative to the document's base URL (not the document's URL)
FAIL An empty-string action content attribute should cause the IDL attribute to return the document's URL (not the document's base URL)assert_equals: expected "http://web-platform.test:8001/html/semantics/forms/the-form-element/form-action-reflection-with-base-url.html" but got "http://web-platform.test:8001/common/blank.html"
FAIL A missing action content attribute should cause the IDL attribute to return the document's URL (not the document's base URL)assert_equals: expected "http://web-platform.test:8001/html/semantics/forms/the-form-element/form-action-reflection-with-base-url.html" but got ""
Harness: the test ran to completion.
